Create 2.py
A young boy was sent by his mother to get some eggs from a nearby supermarket for his birthday party.
On his way home, he was knocked down by a motorist and had all the eggs broken. The boy was unhurt and the motorist agreed to pay for the eggs.
The boy could not remember the exact number of eggs, only that when he took them out in pairs there was one left; similarly there was one left 
when he took them out threes, fours, fives, or sixes at a time. When he took them out in sevens at a time they came out even, that is there was nothing left.
You are required to write a computer solution to determine the smallest number of eggs the boy 
could have had and the total cost if the boy claims that he paid 4 pesewas per egg